backend/requirements.txt
# Web
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
python-dotenv==1.0.1

# LLM + Orchestration
openai>=1.30.0
langgraph>=0.2.0
langchain-core>=0.2.0

# Auth
PyJWT>=2.8.0

# Database
supabase>=2.4.0
redis>=5.0.0
chromadb>=0.5.0

# Workers / Scheduling
apscheduler>=3.10.0

# Eval (optional)
langsmith>=0.1.0
